move TrackerHub classes to Hubs/TrackerHub
